+++ allennlp_models/lm/modules/language_model_heads/gpt2.py
@@ -1,5 +1,5 @@
 from overrides import overrides
-from transformers import GPT2Config, GPT2LMHeadModel
+from transformers.models.gpt2.modeling_gpt2 import GPT2Config, GPT2LMHeadModel
 import torch
 
 from .language_model_head import LanguageModelHead