+++ allennlp_models/structured_prediction/models/srl_bert.py
@@ -4,7 +4,7 @@ from overrides import overrides
 import torch
 from torch.nn.modules import Linear, Dropout
 import torch.nn.functional as F
-from transformers.modeling_bert import BertModel
+from transformers.models.bert.modeling_bert import BertModel
 
 from allennlp.data import TextFieldTensors, Vocabulary
 from allennlp.models.model import Model
@@ -120,6 +120,7 @@ class SrlBert(Model):
             input_ids=util.get_token_ids_from_text_field_tensors(tokens),
             token_type_ids=verb_indicator,
             attention_mask=mask,
+            return_dict=False,
         )
 
         embedded_text_input = self.embedding_dropout(bert_embeddings)